Fixed bug with refresh token causing bad request error for AppAuth refresh token
